Update shovel UI for AMQP 1.0 shovels
To allow cross protocol shovels to be
created and managed through the UI.
To get the desired layout I had to do a
fair bit of mangling of the sammy parameters
before the PUT request. To allow configuration
of different protocols I had to prefix the
input fields with the protocol and direction
and later strip the prefixes and remove the
unwanted properties.

A better approach to this shoudl be found
if further protocols are implemented.

[#138600475]